third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/load-removes-queued-error-event-expected.txt

This is a testharness.js-based test.
FAIL video error eventassert_array_equals: property 1, expected "loadstart" but got "error"
PASS source error event
Harness: the test ran to completion.
